refactor: Change File Structure
