+++ src/main/java/com/db/crud/course/service/student/StudentServiceImpl.java
@@ -1,4 +1,4 @@
-package com.db.crud.course.service;
+package com.db.crud.course.service.student;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Pageable;
